separate merge algorithm
